+++ Source/Shared/Storage/Storage+Support.swift
@@ -2,17 +2,17 @@ import Foundation
 
 public extension Storage {
   func supportData() -> Storage<Data> {
-    let storage = support(transformer: Transformer<Data>.forData())
+    let storage = support(transformer: TransformerFactory.forData())
     return storage
   }
 
   func supportImage() -> Storage<Image> {
-    let storage = support(transformer: Transformer<Image>.forImage())
+    let storage = support(transformer: TransformerFactory.forImage())
     return storage
   }
 
   func supportCodable<U: Codable>(ofType: U.Type) -> Storage<U> {
-    let storage = support(transformer: Transformer<U>.forCodable(ofType: U.self))
+    let storage = support(transformer: TransformerFactory.forCodable(ofType: U.self))
     return storage
   }
 }
