+++ Tests/iOS/Tests/Storage/DiskStorageTests.swift
@@ -5,12 +5,12 @@ final class DiskStorageTests: XCTestCase {
   private let key = "youknownothing"
   private let testObject = User(firstName: "John", lastName: "Snow")
   private let fileManager = FileManager()
-  private var storage: DiskStorage!
+  private var storage: DiskStorage<User>!
   private let config = DiskConfig(name: "Floppy")
 
   override func setUp() {
     super.setUp()
-    storage = try! DiskStorage(config: config)
+    storage = try! DiskStorage<User>(config: config, transformer: Transformer<User>.forCodable(ofType: User.self))
   }
 
   override func tearDown() {
